+++ lib/xcodeproj/gem_version.rb
@@ -1,5 +1,5 @@
 module Xcodeproj
   # The version of the xcodeproj gem.
   #
-  VERSION = '1.8.1'.freeze unless defined? Xcodeproj::VERSION
+  VERSION = '1.8.2'.freeze unless defined? Xcodeproj::VERSION
 end