+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Normalize xcconfig path when generating includes.
   [bclymer](https://github.com/bclymer)
 
+
+
 ## 1.8.1 (2019-02-19)
 
 ##### Enhancements