+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Xcodeproj Changelog
 
-## Master
+## 1.8.2 (2019-03-27)
 
 ##### Enhancements